back/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
drf-spectacular>=0.27
mongoengine>=0.27
pymongo>=4.6
celery[redis]>=5.3
redis>=5.0
django-ratelimit>=4.1
gunicorn>=21.0
